memberships database documentation

memberships database documentation

Usage

memberships

Format

The memberships database is a list that contains the following 2 datasets: COW_MEM, IIGO_MEM. For more information and references to each of the datasets used, please use the data_source() and data_contrast() functions.

COW_MEM:

A dataset with 14317 observations and the following 7 variables: stateID, StateName, igoID, Title, cowigoNR, Beg, End.

IIGO_MEM:

A dataset with 3502 observations and the following 7 variables: stateID, StateName, igoID, Title, Beg, End, Year.

Details

#> $COW_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |           14317|    399    |     2.79      |
#> |   StateName   |  character  |           14317|     0     |       0       |
#> |     igoID     |  character  |           14317|     0     |       0       |
#> |     Title     |  character  |           14317|     0     |       0       |
#> |   cowigoNR    |  character  |           14317|     0     |       0       |
#> |      Beg      |    mdate    |           14317|     0     |       0       |
#> |      End      |    mdate    |           14317|     0     |       0       |
#> ----------------------------------------------------------------------------
#>
#>
#> $IIGO_MEM
#> ----------------------------------------------------------------------------
#> |  Column Name  |  Data Type  |  Observations  |  Missing  |  Missing (%)  |
#> ----------------------------------------------------------------------------
#> |    stateID    |  character  |            3502|    11     |     0.31      |
#> |   StateName   |  character  |            3502|     0     |       0       |
#> |     igoID     |  character  |            3502|     9     |     0.26      |
#> |     Title     |  character  |            3502|     0     |       0       |
#> |      Beg      |    mdate    |            3502|    233    |     6.65      |
#> |      End      |    mdate    |            3502|   2437    |     69.59     |
#> |     Year      |    mdate    |            3502|     0     |       0       |
#> ----------------------------------------------------------------------------